requirements.txt
aiohttp
aiofiles
asyncio
ffmpeg-python
pillow
pymongo[srv]
pyrogram==1.4.16
hachoir
lyricsgenius
py-tgcalls==0.8.6
wget
python-dotenv
youtube-search-python
youtube-search
speedtest-cli
motor
psutil
requests
rich
dnspython
yt-dlp==2021.11.10.1
tgcrypto
gitpython
heroku3
spotipy
bs4
html5lib
numpy
pytube
